# back/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Skillama API",
    "description": "AI Code Generation Server with Anthropic Claude 3",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://skillama.loca.lt"
    }
  ],
  "paths": {
...
    "/api/generate-code": {
      "post": {
        "operationId": "generateCode",
        "summary": "Generate Code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "projectName",
                  "prompt",
                  "filename"
                ],
                "properties": {
                  "projectName": {
                    "type": "string"
                  },
                  "prompt": {
                    "type": "string"
                  },
                  "filename": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully generated code"
          }
        }
      }
    },
...
  }
}